requirements.txt
alabaster==0.7.12
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.5.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.5
awscli==1.19.22
Babel==2.9.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
botocore==1.20.22
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coloredlogs==15.0
colorlover==0.3.0
coverage==5.5
cssselect==1.1.0
cssutils==2.2.0
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.22
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.7.1
descartes==1.1.0
distlib==0.3.1
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.5
GitPython==3.1.14
glyph==2.6.5b0
gunicorn==20.0.4
# hdbscan==0.8.27
humanfriendly==9.1
idna==2.10
imagesize==1.2.0
importlib-metadata==3.7.2
iniconfig==1.1.1
install==1.3.4
ipdb==0.13.6
ipykernel==5.5.0
# ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
micropython-binascii==0.0.2
mistune==0.8.4
mizani==0.7.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.22.0
oauthlib==3.1.0
packaging==20.9
palettable==3.3.0
pandas==1.1.5
pandasql==0.7.3
pandocfilters==1.4.3
parameterized==0.8.1
parso==0.8.1
pathlib==1.0.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
plotly-express==0.4.1
plotnine==0.7.1
pluggy==0.13.1
premailer==3.7.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.5
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyactiveresource==2.2.2
pyarrow==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.7.3
pydeck==0.6.1
pyflakes==2.2.0
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.5.4
Send2Trash==1.5.0
ShopifyAPI==8.4.0
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.23
# -e git+https://github.com/parker84/toronto-crime.git@94002d2cbd1a2220a41486dcfdc3187d8808a1c5#egg=src
statsmodels==0.12.2
streamlit==1.14.0
streamlit-analytics==0.4.1
streamlit-metrics==0.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
tweepy==3.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
virtualenv==20.4.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yagmail==0.14.247
zipp==3.4.1
